more cli work, this can probably be merged but it's not complete

+++ lib/subspace/commands/init.rb
@@ -15,9 +15,11 @@ class Subspace::Commands::Init < Subspace::Commands::Base
     FileUtils.mkdir_p File.join @@dest_dir, "vars"
 
     copy ".gitignore"
+    template "../provision.rb"
     template "ansible.cfg"
     template "hosts"
     template "group_vars/all"
+
     create_vault_pass
     environments.each do |env|
       @env = env
@@ -54,11 +56,6 @@ class Subspace::Commands::Init < Subspace::Commands::Base
     File.basename(Dir.pwd)
   end
 
-  def copy(src, dest = nil)
-    dest ||= src
-    FileUtils.cp File.join(@@provision_templatedir, src), File.join(@@dest_dir, dest)
-  end
-
   def environments
     %w(production dev)
   end
@@ -68,7 +65,11 @@ class Subspace::Commands::Init < Subspace::Commands::Base
   end
 
   def create_vault_pass
-    File.write File.join(@@dest_dir, ".vault_pass"), SecureRandom.base64
+    if File.exist? File.join(@@dest_dir, ".vault_pass")
+      say ".vault_pass already exists.  Skipping..."
+    else
+      File.write File.join(@@dest_dir, ".vault_pass"), SecureRandom.base64
+    end
   end
 
   def create_vars_file_for_env(env)
